% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catUTF8.R
\name{catUTF8}
\alias{catUTF8}
\title{Print the UTF-8 codes of a string.}
\usage{
catUTF8(string, file = "")
}
\arguments{
\item{string}{A character vector.}

\item{file}{A \code{\link{connection}}, or a character string naming the file to print to. If "" (the default), cat prints to the standard output connection, the console unless redirected by \code{\link{sink}}.}
}
\value{
No results.
}
\description{
Print the UTF-8 codes of a string.
}
\examples{
catUTF8("hello")
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
