% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDTM.R
\name{createDTM}
\alias{createDTM}
\alias{createTDM}
\title{Create a Chinese term-document matrix or a document-term matrix.}
\usage{
createDTM(string, tokenize = NULL, removePunctuation = TRUE,
  removeStopwords = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{tokenize}{A tokenizers function.}

\item{removePunctuation}{Whether to remove the punctuations.}

\item{removeStopwords}{Whether to remove the stop words.}
}
\value{
An object of class \code{TermDocumentMatrix} or class \code{DocumentTermMatrix}.
}
\description{
Create a Chinese term-document matrix or a document-term matrix.
}
\details{
Package "tm" is required.
}
\examples{
\dontrun{
library(tmcn)
data(INTRO)
vec1 <- segmentCN(INTRO, returnType = "tm")

dtm1 <- createDTM(vec1)
inspect(dtm1)

tdm1 <- createTDM(vec1)
inspect(tdm1)
}

}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{string}
