% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHashmapEnv.R
\name{createHashmapEnv}
\alias{createHashmapEnv}
\title{Create an environment for hash mapping.}
\usage{
createHashmapEnv(key, value)
}
\arguments{
\item{key}{A character vector of searching key.}

\item{value}{A character vector of searching value.}
}
\value{
An environment.
}
\description{
Create an environment for hash mapping.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
