% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordFreq.R
\name{createWordFreq}
\alias{createWordFreq}
\title{Create a word frequency data.frame.}
\usage{
createWordFreq(obj, onlyCN = TRUE, nosymbol = TRUE, stopwords = NULL,
  useStopDic = FALSE)
}
\arguments{
\item{obj}{A character vector or \code{DocumentTermMatrix} to calculate words frequency.}

\item{onlyCN}{Whether to keep only Chinese words.}

\item{nosymbol}{Whether to keep symbols.}

\item{stopwords}{A character vector of stop words.}

\item{useStopDic}{Whether to use the default stop words.}
}
\value{
A data.frame.
}
\description{
Create a word frequency data.frame.
}
\examples{
createWordFreq(c("a", "a", "b", "c"), onlyCN = FALSE, nosymbol = TRUE, useStopDic = FALSE)

}
\author{
Jian Li <\email{rweibo@sina.com}>
}
