\name{Dataset.GBK}
\alias{Dataset.GBK}
\docType{data}
\title{GBK character set}
\description{
GBK character set including some useful information.
}
\usage{data(Dataset.GBK)}
\format{
  A data frame with 3 columns.
  \describe{
    \item{\code{GBK}}{Chinese characters in UTF-8}
    \item{\code{py0}}{Unique Pinyin of each character}
    \item{\code{py}}{Pinyin string of each character}
    \item{\code{Radical}}{In Chinese, it means 'Bu Shou'.}
    \item{\code{Stroke_Num_Radical}}{In Chinese, it means the number of 'Bi Hua'.}
    \item{\code{Stroke_Order}}{In Chinese, it means 'Bi Shun'.}
    \item{\code{Structure}}{In Chinese, it means 'Zi Ti Jie Gou'.}
  }
}

\keyword{datasets}
