\name{Dataset.Sim2Tra}
\alias{Dataset.Sim2Tra}
\docType{data}
\title{Dictionary of simplified and traditional Chinese}
\description{
Dictionary of simplified and traditional Chinese
}
\usage{data(Dataset.Sim2Tra)}
\format{
  A data frame with 2 columns.
  \describe{
    \item{\code{Sim}}{a simplified Chinese string}
    \item{\code{Tra}}{a traditional Chinese string}
  }
}

\examples{
data(Dataset.Sim2Tra, envir = .tmcnEnv)
}
\keyword{datasets}
