% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharset.R
\name{getCharset}
\alias{getCharset}
\title{Get the current encoding of the locale.}
\usage{
getCharset()
}
\value{
Character of encoding.
}
\description{
Get the current encoding of the locale.
}
\examples{
getCharset()
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
