\name{getWordFreq}
\alias{getWordFreq}
\title{Get the word frequency data.frame.}
\usage{
  getWordFreq(string, onlyCN = TRUE)
}
\arguments{
  \item{string}{A character vector.}

  \item{onlyCN}{Keep only chinese words.}
}
\value{
  A data.frame.
}
\description{
  Get the word frequency data.frame.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

