% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertWords.R
\name{insertWords}
\alias{insertWords}
\title{Insert new words into dictionary.}
\usage{
insertWords(strwords, package = c("jiebaR", "Rwordseg"))
}
\arguments{
\item{strwords}{Vector of words.}

\item{package}{Use which package, "jiebaR" or "Rwordseg"?}
}
\value{
No results.
}
\description{
Insert new words into dictionary.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
