\name{isGB18030}
\alias{isGB18030}
\title{Indicate whether the encoding of input string is GB18030.}
\usage{
isGB18030(string, combine = FALSE)
}
\arguments{
  \item{string}{A character vector.}

  \item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is GB18030.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

