\name{isGB2312}
\alias{isGB2312}
\title{Indicate whether the encoding of input string is GB2312.}
\usage{
  isGB2312(string)
}
\arguments{
  \item{string}{A character vector.}
}
\value{
  Logical value.
}
\description{
  Indicate whether the encoding of input string is GB2312.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

