\name{isGBK}
\alias{isGBK}
\title{Indicate whether the encoding of input string is GBK.}
\usage{
  isGBK(string)
}
\arguments{
  \item{string}{A character vector.}
}
\value{
  Logical value.
}
\description{
  Indicate whether the encoding of input string is GBK.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

