\name{isUTF8}
\alias{isUTF8}
\title{Indicate whether the encoding of input string is UTF-8.}
\usage{
  isUTF8(string)
}
\arguments{
  \item{string}{A character vector.}
}
\value{
  Logical value.
}
\description{
  Indicate whether the encoding of input string is UTF-8.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

