% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\alias{right}
\title{Extract the left or right substrings in a character vector.}
\usage{
left(string, n)
right(string, n)
}
\arguments{
\item{string}{A character vector.}

\item{n}{How many characters.}
}
\value{
A character vector.
}
\description{
Extract the left or right substrings in a character vector.
}
\examples{
left("hello", 3)
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
