\name{NTUSD}
\alias{NTUSD}
\docType{data}
\title{National Taiwan University Semantic Dictionary}
\description{
National Taiwan University Semantic Dictionary.
}
\usage{data(NTUSD)}
\format{
  A list with 4 components.
  \describe{
    \item{\code{positive_chs}}{Positive words in simplified Chinese}
    \item{\code{negative_chs}}{Negative words in simplified Chinese}
    \item{\code{positive_cht}}{Positive words in traditional Chinese}
    \item{\code{negative_cht}}{Negative words in traditional Chinese}
  }
}
\references{
\url{http://nlg18.csie.ntu.edu.tw:8080/opinion/pub1.html }
}
\keyword{datasets}
