% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revUTF8.R
\name{revUTF8}
\alias{revUTF8}
\title{Revert UTF-8 string to Chinese character.}
\usage{
revUTF8(string, utype = "R")
}
\arguments{
\item{string}{A character vector.}

\item{utype}{UTF-8 string type, the default is R type, such as "<U+XXXX>".}
}
\value{
A character vector.
}
\description{
Revert UTF-8 string to Chinese character.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
