% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentCN.R
\name{segmentCN}
\alias{segmentCN}
\title{Sengment a sentence.}
\usage{
segmentCN(strwords, package = c("Rwordseg", "jiebaR"), nature = FALSE,
  nosymbol = TRUE, returnType = c("vector", "tm"), ...)
}
\arguments{
\item{strwords}{A Chinese sentence in UTF-8.}

\item{package}{Use which package, "Rwordseg" or "jiebaR"?}

\item{nature}{Whether to recognise the nature of the words.}

\item{nosymbol}{Whether to keep symbols in the sentence.}

\item{returnType}{Default is a string vector but we also can choose 'tm' 
to output a single string separated by space so that it can be used by \code{\link[tm]{Corpus}} directly.}

\item{...}{Other arguments of Rwordseg.}
}
\value{
a vector of words (list if input is vecter) which have been segmented or the path of output file.
}
\description{
A function segment Chinese sentence into words.
}
\details{
Either package "Rwordseg" (\url{http://jianl.org/cn/R/Rwordseg.html}) or "jiebaR" (\url{http://qinwenfeng.com/jiebaR}) is required.
}
\examples{
\dontrun{
segmentCN("hello world!")
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
