% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setchs.R
\name{setchs}
\alias{setchs}
\alias{setcht}
\alias{setuk}
\title{Set locale to Simplified Chinese/Traditional Chinese/UK.}
\usage{
setchs(rev = FALSE)
setcht(rev = FALSE)
setuk(rev = FALSE)
}
\arguments{
\item{rev}{Whethet to set the locale back.}
}
\value{
No results.
}
\description{
Set locale to Simplified Chinese/Traditional Chinese/UK.
}
\examples{
setchs()
setchs(rev = TRUE)
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
