\name{setCN}
\alias{setCN}
\title{Set locale to Simplified Chinese.}
\usage{
  setCN(rev = FALSE)
}
\arguments{
  \item{rev}{Whethet to set the locale back.}
}
\value{
  No results.
}
\description{
  Set locale to Simplified Chinese.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

