% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setchs.R
\name{setuk}
\alias{setuk}
\title{Set locale to UK English.}
\usage{
setuk(rev = FALSE)
}
\arguments{
\item{rev}{Whethet to set the locale back.}
}
\value{
No results.
}
\description{
Set locale to UK English.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
