\name{SIMTRA}
\alias{SIMTRA}
\docType{data}
\title{Dictionary of simplified and traditional Chinese}
\description{
Dictionary of simplified and traditional Chinese
}
\usage{data(SIMTRA)}
\format{
  A data frame with 2 columns.
  \describe{
    \item{\code{Sim}}{a simplified Chinese string}
    \item{\code{Tra}}{a traditional Chinese string}
  }
}

\keyword{datasets}
