\name{SPORT}
\alias{SPORT}
\docType{data}
\title{Sport news.}
\description{
Sport news.
}
\usage{data(SPORT)}
\format{
  A data frame with 6 columns.
  \describe{
    \item{\code{id}}{ID of the news.}
    \item{\code{time}}{Time of the news.}
    \item{\code{title}}{Title of the news.}
    \item{\code{class}}{Class of the news, 'B' means Basketball, 'F' means Football.}
    \item{\code{abstract}}{Abstract of the news.}
    \item{\code{content}}{Content of the news.}
  }
}

\keyword{datasets}
