\name{STOPWORDS}
\alias{STOPWORDS}
\docType{data}
\title{Dictionary of Chinese stop words}
\description{
Dictionary of Chinese stop words.
}
\usage{data(STOPWORDS)}
\format{
  A data frame with 1 column.
  \describe{
    \item{\code{word}}{a string vertor of the stop words.}
  }
}

\keyword{datasets}
