% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsCN.R
\name{stopwordsCN}
\alias{stopwordsCN}
\title{Return Chinese stop words.}
\usage{
stopwordsCN(stopwords = NULL, useStopDic = TRUE)
}
\arguments{
\item{stopwords}{A character vector of stop words.}

\item{useStopDic}{Whether to use the default stop words.}
}
\value{
A vector of stop words.
}
\description{
Return Chinese stop words.
}
\examples{
stopwordsCN("yes", useStopDic = FALSE)
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
