% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strstrip.R
\name{strstrip}
\alias{strstrip}
\title{Trim space of a string.}
\usage{
strstrip(string, side = c("both", "left", "right"))
}
\arguments{
\item{string}{A character vector.}

\item{side}{Which side of the string to be trimed, 'both', 'left' or 'right'.}
}
\value{
Trimed vector.
}
\description{
Trim space of a string.
}
\examples{
strstrip(c("\\taaaa ", " bbbb    "))


}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{string}
