\name{toCN}
\alias{toCN}
\title{Convert encoding for right displaying in Chinese.}
\usage{
  toCN(string)
}
\arguments{
  \item{string}{A character vector.}
}
\value{
  Converted vectors.
}
\description{
  Convert encoding for right displaying in Chinese.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

