% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPinyin.R
\name{toPinyin}
\alias{toPinyin}
\title{Convert a chinese text to pinyin format.}
\usage{
toPinyin(string, capitalize = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{capitalize}{Whether to capitalize the first letter of each word.}
}
\value{
A character vector in pinyin format.
}
\description{
Convert a chinese text to pinyin format.
}
\examples{
toPinyin("the quick red fox jumps over the lazy brown dog")


}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\keyword{string}
