% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordFreq.R
\name{createWordFreq}
\alias{createWordFreq}
\title{Create a word frequency data.frame.}
\usage{
createWordFreq(obj, onlyCN = TRUE, stopwords = NULL, useStopDic = TRUE)
}
\arguments{
\item{onlyCN}{Keep only chinese words.}

\item{stopwords}{A character vector of stop words.}

\item{useStopDic}{Whether to use the default stop words.}

\item{string}{A character vector to calculate words frequency.}
}
\value{
A data.frame.
}
\description{
Create a word frequency data.frame.
}
\examples{
createWordFreq(c("a", "a", "b", "c"), onlyCN = FALSE, useStopDic = FALSE)

}
\author{
Jian Li <\email{rweibo@sina.com}>
}
