\name{stopwordsCN}
\alias{stopwordsCN}
\title{Return Chinese stop words.}
\usage{
stopwordsCN(stopwords = NULL, useStopDic = TRUE)
}
\arguments{
  \item{stopwords}{A character vector of stop words.}

  \item{useStopDic}{Whether to use the default stop words.}
}
\value{
A vector of stop words.
}
\description{
Return Chinese stop words.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

