\name{toTrad}
\alias{toTrad}
\title{Convert a Chinese text from simplified to traditional characters and vice versa.}
\usage{
toTrad(string, rev = FALSE)
}
\arguments{
  \item{string}{A Chinese string vector.}

  \item{rev}{Reverse. TRUE means traditional to simplified.
  Default is FALSE.}
}
\value{
Converted vectors.
}
\description{
Convert a chinese text from simplified to traditional
characters and vice versa.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

